+++ crates/devela/src/media/visual/draw/canvas/_.rs
@@ -1,0 +1,20 @@
+//
+//!
+//
+
+crate::mods_in! {
+    mod canvas;
+    #[cfg(feature = "image")]
+    mod raster;
+}
+crate::mods_out! { // _mods
+    _mods {
+        pub use super::{
+            canvas::{Canvas, CanvasRead, CanvasTextel},
+        };
+        #[cfg(feature = "image")]
+        pub use super::{
+            raster::{CanvasRaster, CanvasRasterExt},
+        };
+    }
+}