+++ crates/devela/src/sys/device/display/x11/runtime/_.rs
@@ -1,12 +1,24 @@
+//
+
 crate::mods_in! {
-    mod frontend; // (XBackend), (XFrameCtx), XFrontend
-    mod present; // XPresent, (XPresenter), XRasterRender
+    mod frontend;
+    mod present;
+    #[cfg(feature = "image")]
+    mod raster;
 }
-crate::mods_out! { // _mods
+crate::mods_out! { // _mods, _crate_internals
     _mods {
         pub use super::{
-            frontend::*,
-            present::*,
+            frontend::XFrontend,
+            present::XPresent,
+        };
+        #[cfg(feature = "image")]
+        pub use super::raster::XRasterRenderer;
+    }
+    _crate_internals {
+        pub(crate) use super::{
+            frontend::XFrameCtx, // XBackend
+            present::XPresenter,
         };
     }
 }